BackendService_Ver1.0/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pydantic==2.5.0
python-dotenv==1.0.0
alembic==1.12.1
python-multipart==0.0.6